pyproject.toml
[tool.poetry]
name = "zhackathon"
version = "0.1.0-alpha.1"
description = ""

authors = []
readme = "README.md"
license = ""

[tool.poetry.dependencies]
python = "~3.10"
djangorestframework = "^3.14"
django = "^4.1.3"
django-filter = "^22.1"
drf-spectacular = ">=0.21.1,<0.22.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
pylama = "^8.4.1"
pylint = "^2.15.5"
pylint-django = "^2.5.0"
isort = "^5.10.1"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
exclude = '''
(
    /(
        \.git
        |\.tox
        |migrations
    )/
)
'''
include = '\.pyi?$'
line-length = 120

[tool.isort]
profile = "black"